Player select + IA in progress

+++ BigSwordRPG-C#/Game/Character.cs
@@ -13,7 +13,7 @@ namespace BigSwordRPG.Game
         private float _healthMultiplier;
         private float _attMultiplier;
         private float _healMultiplier;
-        private int _speed;
+        private float _speed;
         private bool _isDead;
 
         private Dictionary<string, Abilities> _abilities;
@@ -32,7 +32,7 @@ namespace BigSwordRPG.Game
         }
         public int Level { get => _level; set => _level = value; }
 
-        public int Speed { get => _speed; set => _speed = value; }
+        public float Speed { get => _speed; set => _speed = value; }
         public bool IsDead { get => _isDead; set => _isDead = value; }
         public Dictionary<string,Abilities> CAbilities { get => _abilities; set => _abilities = value; }
         public int MaxHealth { get => _maxHealth; set => _maxHealth = value; }
@@ -43,7 +43,7 @@ namespace BigSwordRPG.Game
 
 
         //Méthodes
-        public Character(string name, int health, int maxHealth, int level, float healthMultiplier, float attMultiplier, float healMultiplier, int speed, Dictionary<string,Abilities> abilities,bool isDead)
+        public Character(string name, int health, int maxHealth, int level, float healthMultiplier, float attMultiplier, float healMultiplier, float speed, Dictionary<string,Abilities> abilities,bool isDead)
         {
             _name = name;
             _health = health;
